+++ api/lib/application/authentication/index.js
@@ -1,44 +1,11 @@
 import Joi from 'joi';
 import { sendJsonApiError, BadRequestError } from '../http-errors.js';
 import { authenticationController as AuthenticationController } from './authentication-controller.js';
-import { securityPreHandlers } from '../security-pre-handlers.js';
 import { responseAuthenticationDoc } from '../../infrastructure/open-api-doc/authentication/response-authentication-doc.js';
 import { responseObjectErrorDoc } from '../../infrastructure/open-api-doc/livret-scolaire/response-object-error-doc.js';
 
 const register = async function (server) {
   server.route([
-    {
-      method: 'POST',
-      path: '/api/token',
-      config: {
-        auth: false,
-        payload: {
-          allow: 'application/x-www-form-urlencoded',
-        },
-        validate: {
-          payload: Joi.alternatives().try(
-            Joi.object().required().keys({
-              grant_type: 'password',
-              username: Joi.string().required(),
-              password: Joi.string().required(),
-              scope: Joi.string(),
-            }),
-            Joi.object().required().keys({
-              grant_type: 'refresh_token',
-              refresh_token: Joi.string(),
-              scope: Joi.string(),
-            }),
-          ),
-        },
-        pre: [{ method: securityPreHandlers.checkIfUserIsBlocked }],
-        handler: AuthenticationController.createToken,
-        tags: ['api'],
-        notes: [
-          "Cette route permet d'obtenir un refresh token et access token à partir d'un couple identifiant / mot de passe" +
-            " ou un access token à partir d'un refresh token valide.",
-        ],
-      },
-    },
     {
       method: 'POST',
       path: '/api/application/token',
@@ -155,5 +122,5 @@ const register = async function (server) {
   ]);
 };
 
-const name = 'authentication-api';
+const name = 'authentication-api-old';
 export { register, name };
