+++ api/server.js
@@ -25,6 +25,7 @@ import { learnerListRoutes } from './src/prescription/organization-learner/route
 import { prescriberManagementRoutes } from './src/shared/prescriber-management/routes.js';
 import { devcompRoutes } from './src/devcomp/routes.js';
 import { schoolRoutes } from './src/school/routes.js';
+import { authenticationRoutes } from './src/authentication/application/routes.js';
 import { scenarioSimulatorRoutes } from './src/certification/flash-certification/routes.js';
 
 const certificationRoutes = [certificationSessionRoutes, attachTargetProfileRoutes, complementaryCertificationRoutes];
@@ -124,6 +125,7 @@ const setupRoutesAndPlugins = async function (server) {
   const configuration = [].concat(
     plugins,
     routes,
+    authenticationRoutes,
     evaluationRoutes,
     scenarioSimulatorRoutes,
     devcompRoutes,
